feat: replace head with generateMetadata for tickets handle
