+++ package.json
@@ -23,7 +23,7 @@
   },
   "dependencies": {
     "@next-auth/prisma-adapter": "^1.0.7",
-    "@prisma/client": "^4.9.0",
+    "@prisma/client": "^5.3.1",
     "@rive-app/react-canvas": "^4.3.3",
     "@tailwindcss/typography": "github:tailwindcss/typography",
     "@vercel/og": "^0.5.17",
@@ -44,7 +44,7 @@
     "three": "^0.150.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.7.1",
+    "@commitlint/cli": "^17.7.2",
     "@commitlint/config-conventional": "^17.7.0",
     "@next/eslint-plugin-next": "^13.5.3",
     "@svgr/webpack": "^8.1.0",
@@ -59,11 +59,11 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "husky": "^8.0.3",
     "lint-staged": "^14.0.1",
-    "postcss": "^8.4.30",
+    "postcss": "^8.4.31",
     "postcss-import": "^15.1.0",
     "prettier": "^3.0.3",
     "prettier-plugin-tailwindcss": "^0.5.4",
-    "prisma": "^4.9.0",
+    "prisma": "^5.3.1",
     "svgo-loader": "^4.0.0",
     "tailwindcss": "^3.3.3",
     "url-loader": "^4.1.1"
