+++ src/app/tickets/[handle]/page.jsx
@@ -3,8 +3,13 @@ import { notFound } from 'next/navigation';
 import DynamicTicket from 'components/pages/developer-days/dynamic-ticket';
 import Button from 'components/shared/button';
 import Layout from 'components/shared/layout';
+import SEO_DATA from 'constants/seo-data';
 import LiveIcon from 'icons/live.inline.svg';
+import getMetadata from 'utils/get-metadata';
 import prisma from 'utils/prisma';
+
+const buildOgImageUrl = (data) =>
+  data ? '/api/og?'.concat(new URLSearchParams(data)) : '/api/og?';
 
 const TicketPage = async ({ params }) => {
   // eslint-disable-next-line no-use-before-define
@@ -73,6 +78,37 @@ async function getTicketData(handle) {
   }
 
   return userData;
+}
+
+export async function generateMetadata({ params }) {
+  const { handle } = params;
+  let userData;
+
+  if (handle) {
+    try {
+      userData = await prisma.user.findUnique({
+        where: {
+          login: handle,
+        },
+        select: {
+          name: true,
+          email: true,
+          login: true,
+          image: true,
+          id: true,
+          colorSchema: true,
+        },
+      });
+    } catch (err) {
+      console.log('Ticket page head query err', err);
+    }
+  }
+
+  if (userData) {
+    return getMetadata({ ...SEO_DATA.ticket(userData), imagePath: buildOgImageUrl(userData) });
+  }
+
+  return getMetadata({ ...SEO_DATA['404-ticket'] });
 }
 
 export async function generateStaticParams() {
